# apps/be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/boards/{id}": {
      "get": {
        "operationId": "Boards_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        },
        "summary": "Get board by id",
        "tags": [
          "Boards"
        ]
      },
      "patch": {
        "operationId": "Boards_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        },
        "summary": "Update board (name, snap)",
        "tags": [
          "Boards"
        ]
      },
      "delete": {
        "operationId": "Boards_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Soft-delete board",
        "tags": [
          "Boards"
        ]
      }
    },
...
    "/boards/{boardId}/connections": {
      "get": {
        "operationId": "BoardNodeConnectionsList_findAll",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromNodeSocketId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toNodeSocketId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBoardNodeConnections"
                }
              }
            }
          }
        },
        "summary": "List connections for a board (paginated)",
        "tags": [
          "Boards"
        ]
      }
    },
    "/boards/{boardId}/props": {
      "get": {
        "operationId": "BoardNodePropsList_findAll",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nodeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "catalogNodePropertyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBoardNodeProps"
                }
              }
            }
          }
        },
        "summary": "List board node props for a board (paginated)",
        "tags": [
          "Boards"
        ]
      }
    },
...
    "/board-nodes/{id}": {
      "get": {
        "operationId": "BoardNodes_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNode"
                }
              }
            }
          }
        },
        "summary": "Get board node by id",
        "tags": [
          "Board - nodes"
        ]
      },
      "patch": {
        "operationId": "BoardNodes_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNode"
                }
              }
            }
          }
        },
        "summary": "Update board node",
        "tags": [
          "Board - nodes"
        ]
      },
      "delete": {
        "operationId": "BoardNodes_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Soft-delete board node",
        "tags": [
          "Board - nodes"
        ]
      }
    },
    "/board-node-connections": {
      "post": {
        "operationId": "BoardNodeConnections_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardNodeConnectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNodeConnection"
                }
              }
            }
          }
        },
        "summary": "Create connection between two board nodes",
        "tags": [
          "Board - connections"
        ]
      }
    },
    "/board-node-connections/{id}": {
      "get": {
        "operationId": "BoardNodeConnections_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNodeConnection"
                }
              }
            }
          }
        },
        "summary": "Get connection by id",
        "tags": [
          "Board - connections"
        ]
      },
      "patch": {
        "operationId": "BoardNodeConnections_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardNodeConnectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNodeConnection"
                }
              }
            }
          }
        },
        "summary": "Update connection",
        "tags": [
          "Board - connections"
        ]
      },
      "delete": {
        "operationId": "BoardNodeConnections_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Soft-delete connection",
        "tags": [
          "Board - connections"
        ]
      }
    },
    "/board-node-props": {
      "post": {
        "operationId": "BoardNodeProps_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardNodePropDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNodeProp"
                }
              }
            }
          }
        },
        "summary": "Set or create runtime prop value for a board node",
        "tags": [
          "Board - node props"
        ]
      }
    },
    "/board-node-props/{id}": {
      "get": {
        "operationId": "BoardNodeProps_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNodeProp"
                }
              }
            }
          }
        },
        "summary": "Get board node prop by id",
        "tags": [
          "Board - node props"
        ]
      },
      "patch": {
        "operationId": "BoardNodeProps_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardNodePropDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardNodeProp"
                }
              }
            }
          }
        },
        "summary": "Update board node prop value",
        "tags": [
          "Board - node props"
        ]
      },
      "delete": {
        "operationId": "BoardNodeProps_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Soft-delete board node prop",
        "tags": [
          "Board - node props"
        ]
      }
    }
  },
  "info": {
    "title": "Shake API",
    "description": "REST API for the **node catalog** (toolbox definitions) and **board graph** (runtime nodes, connections, props).\n\nEntity IDs are `bigint` in PostgreSQL and serialized as **strings** in JSON.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "number",
            "example": 142,
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "example": 8,
            "minimum": 0
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "Board": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "My board",
            "maxLength": 255
          },
          "snap": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "name"
        ]
      },
...
      "UpdateBoardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "My board"
          },
          "snap": {
            "type": "object",
            "nullable": true,
            "description": "Presentation JSON (viewport, layout); optional",
            "example": {
              "zoom": 1,
              "x": 0,
              "y": 0
            }
          }
        }
      },
...
      "BoardNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "value": {
            "type": "object",
            "example": "default payload",
            "maxLength": 512,
            "nullable": true
          },
          "sockets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardNodeSocket"
            }
          },
          "catalogNodeVersion": {
            "$ref": "#/components/schemas/CatalogNodeVersion"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "boardId",
          "catalogNodeVersionId"
        ]
      },
...
      "BoardNodeConnection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "fromNodeSocketId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "toNodeSocketId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "order": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "boardId",
          "fromNodeSocketId",
          "toNodeSocketId",
          "order"
        ]
      },
      "PaginatedBoardNodeConnections": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardNodeConnection"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "BoardNodeProp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "nodeId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "catalogNodePropertyId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "value": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "boardId",
          "nodeId",
          "catalogNodePropertyId"
        ]
      },
      "PaginatedBoardNodeProps": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardNodeProp"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "UpdateBoardNodeDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "FK to boards.id"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "Pinned catalog version; FK to catalog_node_versions.id"
          },
          "value": {
            "type": "string",
            "nullable": true,
            "maxLength": 512,
            "example": "instance-meta",
            "description": "Optional runtime value on the placed node"
          }
        }
      },
      "CreateBoardNodeConnectionDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "FK to boards.id"
          },
          "fromNodeSocketId": {
            "type": "string",
            "example": "100",
            "description": "FK to board_node_sockets.id"
          },
          "toNodeSocketId": {
            "type": "string",
            "example": "101",
            "description": "FK to board_node_sockets.id"
          },
          "order": {
            "type": "number",
            "example": 0,
            "description": "Ordering for parallel edges"
          }
        },
        "required": [
          "boardId",
          "fromNodeSocketId",
          "toNodeSocketId"
        ]
      },
      "UpdateBoardNodeConnectionDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "FK to boards.id"
          },
          "fromNodeSocketId": {
            "type": "string",
            "example": "100",
            "description": "FK to board_node_sockets.id"
          },
          "toNodeSocketId": {
            "type": "string",
            "example": "101",
            "description": "FK to board_node_sockets.id"
          },
          "order": {
            "type": "number",
            "example": 0,
            "description": "Ordering for parallel edges"
          }
        }
      },
      "CreateBoardNodePropDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "FK to boards.id"
          },
          "nodeId": {
            "type": "string",
            "example": "10",
            "description": "FK to board_nodes.id"
          },
          "catalogNodePropertyId": {
            "type": "string",
            "example": "20",
            "description": "FK to catalog_node_properties.id"
          },
          "value": {
            "type": "object",
            "description": "JSON value per catalog property type",
            "example": 42
          }
        },
        "required": [
          "boardId",
          "nodeId",
          "catalogNodePropertyId"
        ]
      },
      "UpdateBoardNodePropDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "string",
            "example": "1",
            "description": "FK to boards.id"
          },
          "nodeId": {
            "type": "string",
            "example": "10",
            "description": "FK to board_nodes.id"
          },
          "catalogNodePropertyId": {
            "type": "string",
            "example": "20",
            "description": "FK to catalog_node_properties.id"
          },
          "value": {
            "type": "object",
            "description": "JSON value per catalog property type",
            "example": 42
          }
        }
      },
...
    }
  }
}